feat: add makefile for building CLI
