+++ .gitignore
@@ -1,5 +1,4 @@
 /.vscode
-/cmd/storage
 /storage
 *.car
 build
